LAGraph: Avoid duplicate definition of global variables.
Avoid linker issue about duplicate symbols when a binary is linked to
the static lagraph and the static lagraphx libraries.

That is caused by the fact that both libraries define a global variable
`timings`. That variable doesn't need to be global anywhere. So, make it
its scope file static in both places where it is used.